Updated bluray.com
It now searches for movie/tv show titles instead of media releases so it's "cleaner" this way, otherwise you would get a lot of blu-ray releases instead of getting the Movie or TV Show page.

Had to "tweak" the search parameters a little or TV Shows weren't gonna be able to be searched for, and because of this when searching for a movie it may also return its sequels (and prequels to? idk, if a prequel was released after the sequel then it may happen but still, the titles will need to be almost the same which is very unlikely to happen so...) because the search isn't limited with max year of release.

i.e. Searching for the "Now You See Me" movie will return the first and the second movie because the title of the sequel only has a "2" added to its title, and the search parameters only limit the minimum release year which would be 2013 and the sequel (obviously) was released after 2013.